+++ crates/runmat-vm/src/interpreter/dispatch/mod.rs
@@ -1,0 +1,13 @@
+mod calls;
+mod control_flow;
+mod exceptions;
+
+pub use calls::{
+    handle_builtin_outcome, handle_feval_dispatch, output_list_for_user_call, push_single_result,
+    push_user_call_outputs, unpack_prepared_user_call, BuiltinHandling, FevalHandling,
+    PreparedUserDispatch,
+};
+pub use control_flow::{apply_control_flow_action, DispatchDecision};
+pub use exceptions::{
+    parse_exception, prepare_vm_error, redirect_exception_to_catch, ExceptionHandling,
+};
